Factored out preloading.
